+++ Migrations/AppDbContextModelSnapshot.cs
@@ -39,8 +39,8 @@ namespace EMS.Migrations
                         .IsRequired()
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<int>("Phone")
-                        .HasColumnType("int");
+                    b.Property<long>("Phone")
+                        .HasColumnType("bigint");
 
                     b.Property<string>("Role")
                         .IsRequired()
